Rename files JC-674
